Stream read csv file from S3 and log each line for testing
This check in is to capture work in progress that successfully stream reads csv file form minio S3 bucket as part of import process and logs out each line one at a time for development testing.
There is a load of test code needing clean up, which is a TODO for me.