5863 Fix cell formatting with commas in the text
When a CSV value contained a comma as part of it's text the XLS
operation was splitting the row on ',' and creating extra columns.

The solution is to use the Go built in `csv.NewReader` operation to
interpret a csv row/line. This would account for other characters/runes
inside the CSV value.

Resolves: 5863